+++ R/qp-uvb313-temp.r
@@ -20,6 +20,9 @@
 #'   \code{"source_mspct"} object, with member \code{"source_spct"} objects
 #'   named minus5C, plus00C, plus05C, plus10C, plus20C, plus30C, and plus35C.
 #'
+#'   The variables in each member spectrum are as follows: \itemize{ \item
+#'   w.length (nm) \item s.e.irrad (\eqn{\mathrm{W\,m^{-2}\,nm^{-1}}}) }
+#'
 #' @note Instrument used: Macam SR-9010-PC scanning double monochromator
 #' spectroradiometer with a cosine corrected input optics. Recently calibrated.
 #' The lamps were driven by a high frequency electronic ballast.