+++ NEWS.md
@@ -8,6 +8,7 @@ photobiologyLamps 0.5.3
 =======================
 
 * Add data for grow lamps and growth cabinets.
+* Add data for video/photography on-camera LED lamps.
 
 photobiologyLamps 0.5.2
 =======================